Adjusted the primary keys and Ids

+++ src/main/java/com/rcoem/filmrentalstore/entities/City.java
@@ -21,7 +21,7 @@ public class City {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "city_id")
-    private Long cityId;
+    private Short cityId;
 
     @Column(name = "city")
     private String city;
