+++ src/main/java/com/rcoem/filmrentalstore/repository/ActorRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.rcoem.filmrentalstore.entities.Actor;
 
-public interface ActorRepository extends JpaRepository<Actor,Long>{
+public interface ActorRepository extends JpaRepository<Actor,Short>{
 
 }
