+++ src/main/java/com/rcoem/filmrentalstore/repository/PaymentRepository.java
@@ -11,6 +11,6 @@ import java.util.List;
 
 
 @Repository
-public interface PaymentRepository extends JpaRepository<Payment, Long> {
+public interface PaymentRepository extends JpaRepository<Payment, Short> {
     Page<Payment> findByStaff_Username(@Param("username") String username, Pageable pageable);
 }
