+++ src/main/java/com/rcoem/filmrentalstore/repository/RentalRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface RentalRepository extends JpaRepository<Rental, Long> {
+public interface RentalRepository extends JpaRepository<Rental, Integer> {
     @Query(value = """
             SELECT 
                 f.title AS filmName,
