+++ src/main/java/com/rcoem/filmrentalstore/repository/StoreRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface StoreRepository extends JpaRepository<Store,Long> {
+public interface StoreRepository extends JpaRepository<Store,Byte> {
 
     Optional<Store> findByAddress(Address testAddress);
 }
